add middleware

+++ src/modules/course-categories/course-categories.controller.ts
@@ -7,28 +7,43 @@ import {
   Delete,
   Put,
   Patch,
+  Query,
+  HttpException,
+  HttpStatus,
+  Req,
 } from '@nestjs/common';
+import { Request } from 'express';
 import { CourseCategoriesService } from './course-categories.service';
-import { CreateCourseDto } from '../courses/dto/create-course.dto';
 import { UpdateCourseCategoryDto } from './dto/update-course-category.dto';
-@Controller('course_categories')
+import { CreateCourseCategoryDto } from './dto/create-course-category.dto';
+import { Types } from 'mongoose';
+
+@Controller('course-categories')
 export class CourseCategoriesController {
   constructor(
     private readonly courseCategoriesService: CourseCategoriesService,
   ) {}
 
   @Get()
-  findAll() {
-    return this.courseCategoriesService.findAll();
+  findAll(@Query('includes') includes: string, @Req() req: Request) {
+    console.log(req.user);
+    return this.courseCategoriesService.findAll(includes);
   }
 
   @Get('/:id')
-  find(@Param('id') id: string) {
-    return this.courseCategoriesService.find(id);
+  async find(@Param('id') id: string, @Query('includes') includes: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new HttpException('ID không hợp lệ', HttpStatus.BAD_REQUEST);
+    }
+    const res = await this.courseCategoriesService.find(id, includes);
+    if (!res) {
+      throw new HttpException('không tìm thấy id', HttpStatus.NOT_FOUND);
+    }
+    return res;
   }
 
   @Post()
-  create(@Body() createCourseDto: CreateCourseDto) {
+  create(@Body() createCourseDto: CreateCourseCategoryDto) {
     return this.courseCategoriesService.create(createCourseDto);
   }
 
@@ -49,7 +64,14 @@ export class CourseCategoriesController {
   }
 
   @Delete('/:id')
-  delete(@Param('id') id: string) {
-    return this.courseCategoriesService.delete(id);
+  async delete(@Param('id') id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new HttpException('ID không hợp lệ', HttpStatus.BAD_REQUEST);
+    }
+    const res = await this.courseCategoriesService.delete(id);
+    if (!res) {
+      throw new HttpException('không tìm thấy id', HttpStatus.NOT_FOUND);
+    }
+    return res;
   }
 }
